PLAT-338: mobved finalized state to package
