+++ ledger-core/virtual/object/finalizedstate/state_report.go
@@ -3,7 +3,7 @@
 // This material is licensed under the Insolar License version 1.0,
 // available at https://github.com/insolar/assured-ledger/blob/master/LICENSE.md.
 
-package object
+package finalizedstate
 
 import (
 	"context"
@@ -14,7 +14,6 @@ import (
 	"github.com/insolar/assured-ledger/ledger-core/insolar/payload"
 	"github.com/insolar/assured-ledger/ledger-core/network/messagesender"
 	messageSenderAdapter "github.com/insolar/assured-ledger/ledger-core/network/messagesender/adapter"
-	"github.com/insolar/assured-ledger/ledger-core/pulse"
 	"github.com/insolar/assured-ledger/ledger-core/reference"
 	"github.com/insolar/assured-ledger/ledger-core/vanilla/injector"
 	"github.com/insolar/assured-ledger/ledger-core/vanilla/throw"
@@ -56,36 +55,6 @@ func (sm *SMStateReport) GetInitStateFor(smachine.StateMachine) smachine.InitFun
 
 func (sm *SMStateReport) GetStateMachineDeclaration() smachine.StateMachineDeclaration {
 	return sm
-}
-
-type SharedReportAccessor struct {
-	smachine.SharedDataLink
-}
-
-func (v SharedReportAccessor) Prepare(fn func(report payload.VStateReport)) smachine.SharedDataAccessor {
-	return v.PrepareAccess(func(data interface{}) bool {
-		fn(*data.(*payload.VStateReport))
-		return false
-	})
-}
-
-type ReportKey struct {
-	ObjectReference reference.Global
-	Pulse           pulse.Number
-}
-
-func BuildReportKey(object reference.Global, pulse pulse.Number) ReportKey {
-	return ReportKey{
-		ObjectReference: object,
-		Pulse:           pulse,
-	}
-}
-
-func GetSharedStateReport(ctx smachine.InOrderStepContext, object reference.Global, pn pulse.Number) (SharedReportAccessor, bool) {
-	if v := ctx.GetPublishedLink(BuildReportKey(object, pn)); v.IsAssignableTo((*payload.VStateReport)(nil)) {
-		return SharedReportAccessor{v}, true
-	}
-	return SharedReportAccessor{}, false
 }
 
 func (sm *SMStateReport) migrationDefault(ctx smachine.MigrationContext) smachine.StateUpdate {
