+++ ledger-core/virtual/object/finalizedstate/state_report_test.go
@@ -3,7 +3,7 @@
 // This material is licensed under the Insolar License version 1.0,
 // available at https://github.com/insolar/assured-ledger/blob/master/LICENSE.md.
 
-package object
+package finalizedstate
 
 import (
 	"testing"
@@ -11,29 +11,71 @@ import (
 	"github.com/gojuno/minimock/v3"
 	"github.com/stretchr/testify/require"
 
+	"github.com/insolar/assured-ledger/ledger-core/conveyor"
 	"github.com/insolar/assured-ledger/ledger-core/conveyor/smachine"
 	"github.com/insolar/assured-ledger/ledger-core/insolar/payload"
 	"github.com/insolar/assured-ledger/ledger-core/instrumentation/inslogger"
+	"github.com/insolar/assured-ledger/ledger-core/pulse"
 	"github.com/insolar/assured-ledger/ledger-core/reference"
 	"github.com/insolar/assured-ledger/ledger-core/testutils"
+	"github.com/insolar/assured-ledger/ledger-core/testutils/gen"
 	messageSenderWrapper "github.com/insolar/assured-ledger/ledger-core/testutils/messagesender"
+	"github.com/insolar/assured-ledger/ledger-core/vanilla/longbits"
 	"github.com/insolar/assured-ledger/ledger-core/virtual/descriptor"
 )
+
+func buildStateReport(status payload.VStateReport_StateStatus, state descriptor.Object) payload.VStateReport {
+	var (
+		pd          = pulse.NewFirstPulsarData(10, longbits.Bits256{})
+		smObjectID  = gen.UniqueIDWithPulse(pd.PulseNumber)
+		smGlobalRef = reference.NewSelf(smObjectID)
+	)
+
+	res := payload.VStateReport{
+		Callee:          smGlobalRef,
+		ProvidedContent: &payload.VStateReport_ProvidedContentBody{},
+		Status:          status,
+	}
+
+	if state != nil {
+		res.LatestDirtyState = state.HeadRef()
+		class, _ := state.Class()
+		res.ProvidedContent.LatestDirtyState = &payload.ObjectState{
+			Reference: state.StateID(),
+			Parent:    state.Parent(),
+			Class:     class,
+			State:     state.Memory(),
+		}
+	}
+	return res
+}
+
+func newSMReportWithPulse() *SMStateReport {
+	var (
+		pd          = pulse.NewFirstPulsarData(10, longbits.Bits256{})
+		pulseSlot   = conveyor.NewPresentPulseSlot(nil, pd.AsRange())
+		smObjectID  = gen.UniqueIDWithPulse(pd.PulseNumber)
+		smGlobalRef = reference.NewSelf(smObjectID)
+		smReport    = &SMStateReport{
+			SMReport: SMReport{
+				Reference: smGlobalRef,
+			},
+			pulseSlot: &pulseSlot,
+		}
+	)
+
+	return smReport
+}
 
 func TestSMStateReport_SendVStateReport_IfDescriptorSet(t *testing.T) {
 	var (
 		mc = minimock.NewController(t)
 
-		smObject             = newSMObjectWithPulse()
 		msgVStateReportCount = 0
 	)
 
-	smObject.SetDescriptor(descriptor.NewObject(reference.Global{}, reference.Local{}, reference.Global{}, nil, reference.Global{}))
-	smObject.SharedState.SetState(HasState)
-
 	smReport := newSMReportWithPulse()
-	smReport.report = smObject.BuildStateReport()
-	smReport.report.ProvidedContent.LatestDirtyState = smObject.BuildLatestDirtyState()
+	smReport.report = buildStateReport(payload.Ready, descriptor.NewObject(reference.Global{}, reference.Local{}, reference.Global{}, nil, reference.Global{}))
 
 	messageService := messageSenderWrapper.NewServiceMockWrapper(mc)
 	checkMessageFn := func(msg payload.Marshaler) {
@@ -62,13 +104,11 @@ func TestSMStateReport_SendVStateReport_IfDescriptorNotSetAndStateEmpty(t *testi
 	var (
 		mc = minimock.NewController(t)
 
-		smObject             = newSMObjectWithPulse()
 		msgVStateReportCount = 0
 	)
-	smObject.SetState(Empty)
 
 	smReport := newSMReportWithPulse()
-	smReport.report = smObject.BuildStateReport()
+	smReport.report = buildStateReport(payload.Empty, nil)
 
 	messageService := messageSenderWrapper.NewServiceMockWrapper(mc)
 	checkMessageFn := func(msg payload.Marshaler) {
@@ -84,7 +124,6 @@ func TestSMStateReport_SendVStateReport_IfDescriptorNotSetAndStateEmpty(t *testi
 
 	smReport.messageSender = messageSender.Mock()
 
-	require.Equal(t, nil, smObject.descriptor)
 	execCtx := smachine.NewExecutionContextMock(mc).
 		JumpMock.Return(smachine.StateUpdate{}).
 		SetDefaultMigrationMock.Set(testutils.AssertMigration(t, smReport.migrationDefault))
